Refactor calculator into modules and add image printing
