+++ kalkulator/__init__.py
@@ -1,0 +1,5 @@
+"""Pakiet logiki aplikacji Kalkulator Rekruso."""
+
+from .ui import FalaBApp, main
+
+__all__ = ["FalaBApp", "main"]
